requirements/base.txt
noise==1.2.1
pillow==2.6.1
numpy==1.9.1